% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{lfs_convert}
\alias{lfs_convert}
\title{Convert files to Rds for faster compiling in future}
\usage{
lfs_convert(
  directory,
  output_directory,
  filter_files = NULL,
  incremental = FALSE
)
}
\arguments{
\item{directory}{Directory with raw LFS files}

\item{output_directory}{Target directory to save Rds files}

\item{filter_files}{Limit to certain files, using a named vector}

\item{incremental}{Only convert files which don't exist in output directory}
}
\value{
List of output
}
\description{
Converts a directory of .sav files to a new folder with .Rds files
}
\details{
Note: importing labelled from haven is required to import haven when
tests are run - otherwise it can't use the haven labelled format.
}
