% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lfs_extra_mappings}
\alias{lfs_extra_mappings}
\title{Add Custom Column Mappings}
\usage{
lfs_extra_mappings(file_name = "extra_lfs_mappings.R")
}
\arguments{
\item{file_name}{Name of mapping file}
}
\description{
Create file with function that can be used to add columns
}
