% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{lfs_flag_postgraduate}
\alias{lfs_flag_postgraduate}
\title{Add a post-gradaute flag column to LFS data}
\usage{
lfs_flag_postgraduate(lfs, column_name = "POSTGRADUATE")
}
\arguments{
\item{lfs}{LFS dataset}

\item{column_name}{Name of the column - by default POST-GRADUATE}
}
\value{
tibble with extra teacher column
}
\description{
Add a post-gradaute flag column to LFS data
}
