% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{lfs_flag_teacher}
\alias{lfs_flag_teacher}
\title{Add a teacher flag column to LFS data}
\usage{
lfs_flag_teacher(lfs, column_name = "TEACHER")
}
\arguments{
\item{lfs}{LFS dataset}

\item{column_name}{Name of the column - by default TEACHER}
}
\value{
tibble with extra teacher column
}
\description{
Add a teacher flag column to LFS data
}
