% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lfs_load}
\alias{lfs_load}
\title{Load LFS Data}
\usage{
lfs_load(data.table = TRUE)
}
\arguments{
\item{data.table}{Import as data.table?}
}
\value{
A tibble with lfs data
}
\description{
This works only with the \code{DATA_DIRECTORY} environment variable set to a folder.
You can either manually save \code{lfs_data.fst} there, or it will automatically be
saved during compilation if the \code{DATA_DIRECTORY} variable is present while running
\code{lfs_compile()}
}
