% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{lfs_pick_column}
\alias{lfs_pick_column}
\title{Picks column names in order of appearanceA}
\usage{
lfs_pick_column(vector_of_names, column_names)
}
\arguments{
\item{vector_of_names}{character vector of column names}

\item{column_names}{vector of columns from lfs}
}
\value{
First name that appears in cols
}
\description{
When column names change over time, picks the first
one that appears from the list in order of appearance.
If it doesn't appear, returns a character NA.
}
