% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_experimental.R
\name{lfs_summarise_hours}
\alias{lfs_summarise_hours}
\title{Group by and Summarise for hours worked}
\usage{
lfs_summarise_hours(data, ...)
}
\arguments{
\item{data}{LFS Dataset}

\item{...}{Variables to group by}
}
\value{
Tibble of outputs
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
