% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lfs_summarise_unemployment}
\alias{lfs_summarise_unemployment}
\title{Group by and Summarise for ONS unemployment}
\usage{
lfs_summarise_unemployment(lfs, ...)
}
\arguments{
\item{lfs}{LFS Dataset}

\item{...}{Variables to group by}
}
\value{
Tibble or data.table with unemployment data
}
\description{
Group by and Summarise for ONS unemployment
}
