% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lfs_tidy_file}
\alias{lfs_tidy_file}
\title{Selects and renames columns from LFS File}
\usage{
lfs_tidy_file(file, total_files, file_format, extra_mappings = NULL)
}
\arguments{
\item{file}{Index of file}

\item{total_files}{Vector of file names - helper function for cli}

\item{file_format}{File extension: sav csv or rds}

\item{extra_mappings}{Either NULL (use default) or an R script with the lfs_extra_mappings function}
}
\value{
List of output
}
\description{
Checks which variables are present and picks them in order of priority
eg. PWT20 > PWT18 etc. Renames them so that they can be binded to one dataset
}
\details{
To automatically save the compiled file in a directory ready to be loaded with
\code{lfs_load()}, set the \code{DATA_DIRECTORY} environment variable to point at a folder.
}
\keyword{internal}
