% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ons_time_series}
\alias{ons_time_series}
\title{Get time series from ONS Api}
\usage{
ons_time_series(dataset_id, timeseries_id)
}
\arguments{
\item{dataset_id}{Name of ONS dataset}

\item{timeseries_id}{Name of ONS time series}
}
\value{
A Jsonlite(?) object from the ONS API
}
\description{
Get a list with ONS API output
}
\examples{

# CPIH Inflation
ons_time_series("MM23", "L55O") \%>\%
  purrr::pluck("years") \%>\%
  dplyr::select(date, value)

# Unemployment Rate
ons_time_series("LMS", "MGSX") \%>\%
  purrr::pluck("months") \%>\%
  dplyr::select(date, value)
}
\keyword{internal}
