% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylfs-package.R
\docType{package}
\name{tidylfs-package}
\alias{tidylfs}
\alias{tidylfs-package}
\title{tidylfs: Tools for analysing the ONS Labour Force Survey}
\description{
Tools for analysts working with the LFS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wklimowicz/tidylfs}
  \item \url{https://wklimowicz.github.io/tidylfs/}
  \item Report bugs at \url{https://github.com/wklimowicz/tidylfs/issues}
}

}
\author{
\strong{Maintainer}: Wojtek Klimowicz \email{wojtekklimowicz96@gmail.com}

}
\keyword{internal}
