% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_mapping.R
\name{variable_mapping}
\alias{variable_mapping}
\title{Retrieve variable mapping for LFS/APS compilation}
\usage{
variable_mapping(lfs)
}
\arguments{
\item{lfs}{A data.frame containing the LFS/APS data.}
}
\value{
A data.frame with the variable mapping.
}
\description{
Retrieve variable mapping for LFS/APS compilation
}
