% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{aps_load}
\alias{aps_load}
\title{Load APS Data}
\usage{
aps_load(data.table = TRUE)
}
\arguments{
\item{data.table}{Import as data.table?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be removed in a future version of the package.

It is recommended to explicitly save the compiled data to persist it.
}
\keyword{internal}
