% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lfs_compile}
\alias{lfs_compile}
\title{Compile to 1 R file}
\usage{
lfs_compile(
  lfs_directory,
  extra_mappings = NULL,
  save_location = "package",
  save_variables_report = TRUE
)
}
\arguments{
\item{lfs_directory}{Path to directory with LFS files}

\item{extra_mappings}{Either NULL (use default) or a file
which has the custom mapping function.
See \code{\link{lfs_extra_mappings}}}

\item{save_location}{File path for \code{Rds} file to save. By default, saves the
output in the package directory, so it's always accessible with \code{lfs_load}.}

\item{save_variables_report}{Save a csv with the list of picked variables?}
}
\value{
Nothing
}
\description{
Compile to 1 R file
}
