% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lfs_compile}
\alias{lfs_compile}
\title{Compile to single data.frame}
\usage{
lfs_compile(
  directory,
  dataset = c("lfs", "aps"),
  filter_years = NULL,
  extra_mappings = NULL,
  aps = lifecycle::deprecated(),
  save_variables_report = lifecycle::deprecated(),
  save_to_folder = lifecycle::deprecated()
)
}
\arguments{
\item{directory}{Path to directory with LFS files}

\item{dataset}{Either "lfs" or "aps".

"lfs" expects files named like "2004 Q1.sav" or "2004 Q2.sav",

"aps" expects files named like "APS 2004.sav" or "APS 2005.sav".}

\item{filter_years}{Vector of years to filter down to}

\item{extra_mappings}{Either NULL (use default) or a file
which has the custom mapping function.
See
\code{vignette("Adding_Variables", package = "tidylfs")}}

\item{aps}{Deprecated. Use \code{dataset = "aps"} instead.}

\item{save_variables_report}{Deprecated.  Use \code{variables_mapping(lfs)} instead.}

\item{save_to_folder}{Deprecated. Use \code{lfs <- lfs_compile(..)}, and save manually to a folder.}
}
\value{
A data.table with a stacked LFS dataset. Access the variable mappinging with \code{variable_mapping()}.
}
\description{
Compiles seperate fst files into one, picking relevant columns
}
