% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{lfs_flag_graduate_ons}
\alias{lfs_flag_graduate_ons}
\title{Add an ONS graduate flag column to LFS data}
\usage{
lfs_flag_graduate_ons(lfs, column_name = "GRADUATE_ONS")
}
\arguments{
\item{lfs}{LFS dataset}

\item{column_name}{Name of the column - by default GRADUATE_ONS}
}
\value{
tibble with extra teacher column
}
\description{
Add an ONS graduate flag column to LFS data
}
