% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lfs_summarise_salary}
\alias{lfs_summarise_salary}
\title{Group by and Summarise into salary ONS Salary}
\usage{
lfs_summarise_salary(lfs, ...)
}
\arguments{
\item{lfs}{LFS Dataset}

\item{...}{Variables to group by}
}
\value{
Tibble of outputs
}
\description{
Group by and Summarise into salary ONS Salary
}
