% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_experimental.R
\name{lfs_summarise_union}
\alias{lfs_summarise_union}
\title{Group by and Summarise for unionisation rates}
\usage{
lfs_summarise_union(lfs, ...)
}
\arguments{
\item{lfs}{LFS Dataset}

\item{...}{Variables to group by}
}
\value{
Tibble of outputs
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. The code has moved to a vignette and should be used
explicitly in analysis. Refer to the package documentation for detailed usage.
}
\keyword{internal}
